Restructure to use workspace, allow cross-platform build, host-platform build, host-platform test
