+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "rust-analyzer.check.allTargets": false,
+    "rust-analyzer.cargo.features": "all"
+}
