+++ ec-service-lib/src/error.rs
