+++ ec-service-lib/src/thermal/mod.rs
