--- platform/ihv1-sp/.cargo/config.toml
@@ -1,2 +1,0 @@
-[build]
-target = "aarch64-unknown-none"
