+++ platform/ihv1-sp/src/baremetal/panic.rs
