--- platform/qemu-sp/rustfmt.toml
@@ -1,3 +1,0 @@
-group_imports = "StdExternalCrate"
-imports_granularity = "Module"
-max_width = 120
