+++ platform/qemu-sp/src/baremetal/exceptions.rs
