+++ platform/qemu-sp/src/baremetal/panic.rs
