+++ rust-toolchain.toml
@@ -1,3 +1,3 @@
 [toolchain]
 components = ["rust-src", "llvm-tools-preview", "rustfmt", "rust-analyzer"]
-targets = [ "aarch64-unknown-none-softfloat" ]
+targets = ["aarch64-unknown-none-softfloat"]
